refactor zmanim controller and moved out logic to zmanimService
